fieldable panels panes custom widget
This is a work in progress. Started out as a feature that was built
using Panopoly and Fieldable Panels Pane plugin and module. Then it was
converted to a module so it could be included along with all the built
in widgets that add custom content widgets. There are five integration
widgets inclusive to this feature module; facebook page feed, twitter
feed, twitter like button. These widgets allow a content editor to
customize the look and feel of the social media on the their website,
and allow them to save the configuration for later use. TODO: Add forms
for administrator or other roles that can configure the default settings
for the module. Then role in field and field instances using the field
api rather than features base and instance methods for adding fields
(perhaps)